requirements.txt
# Framework
fastapi==0.105.0
pydantic==2.5.3
pydantic-settings==2.1.0
uvicorn==0.25.0

# Database
SQLAlchemy~=2.0.23
alembic~=1.12.1
asyncmy
sqlmodel==0.0.14

# Security
pyjwt==2.3.0
cryptography
python-jose[cryptography]~=3.3.0

# These requirements are only required for development, not for production:

# Testing
coverage~=7.3.2
pytest-cov~=4.1.0
pytest~=7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
aiounittest~=1.4.2
aiosqlite==0.19.0
polyfactory==2.12.0

# Linters
pylint==3.0.2
flake8==6.1.0
mypy==1.7.1
types-PyYAML==6.0.12.12